% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCmeasurement.R
\docType{class}
\name{BCmeasurement-class}
\alias{BCmeasurement-class}
\alias{BCmeasurement}
\title{Structure of the S4 class "BCmeasurement"}
\description{
Structure of the S4 class "BCmeasurement"
}
\section{Slots}{

\describe{
\item{\code{mets}}{data.table with column information (i.e., metabolite info)}

\item{\code{samples}}{data.table with row information (i.e., sample info)}

\item{\code{conc}}{Numeric matrix of measured concentrations. Rows are samples,
columns are metabolites.}

\item{\code{indicators}}{Numeric matrix with concentration-derived metabolism
indicator metrics. (Not yet implemented)}

\item{\code{status}}{Character matrix of each metabolites measurement quality.}
}}

