% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeBC.R
\name{imputeBC}
\alias{imputeBC}
\title{Imputing missing values}
\usage{
imputeBC(
  ms,
  method,
  max.missing = 0.2,
  keep.status = c("Valid", "< LLOQ", "> ULOQ", "Imputed")
)
}
\arguments{
\item{ms}{An S4-object of class \link{BCmeasurement}}

\item{method}{Imputation method. See details.}

\item{max.missing}{Portion of values per metabolite that are in maximum
allowed to be missing in order to be included in imputation.}

\item{keep.status}{Status of values that should be kept as learning data (i.e., not
subject to imputation).}
}
\value{
An object of class \link{BCmeasurement} with imputed values. If a
value was imputed or not is stored in the \code{object@status} slot with the tag
"imputed".
}
\description{
Imputes missing values or values of low quality.
}
\details{
Possible imputation methods are:
\itemize{
\item \if{html}{\out{<b>}}LOD\if{html}{\out{</b>}} - Values with the status "< LOD" are imputed with the respective LOD
of the focal metabolite and Biocrates plate.
\item \code{RF} - Nonparametric Missing Value Imputation using Random Forest as
implemented in \link[missForest]{missForest}. Note that RF imputation in
non-deterministic.
}
}
