% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBC.R
\name{readBC}
\alias{readBC}
\title{Reads an excel file with a Biocrates metabolomics data set}
\usage{
readBC(file, sheet = "Data Export", beautify = TRUE)
}
\arguments{
\item{file}{Path to excel file.}

\item{sheet}{Name of data sheet with main table.}

\item{beautify}{logical. Should specific metabolite names and classes be
renamed? See details. (Default: TRUE)}
}
\value{
An object of class \link{BCmeasurement}.
}
\description{
A function that transforms a excel table with Biocrates results into an
S4 R object that is easier to handle for data processing, QC, imputation,
statistics and data merging.
}
\details{
If TRUE, the parameter 'beautify' renames the following entities:
\itemize{
\item "Amine Oxides" -> "Amine oxides"
\item "Aminoacids" -> "Amino acids"
\item "Aminoacids Related" -> "Amino acid-related"
\item "Bile Acids" -> "Bile acids"
\item "Biogenic Amines" -> "Biogenic amines"
\item "Carboxylic Acids" -> "Carboxylic acids"
\item "Cholesterol Esters" -> "Cholesterol esters"
\item "Fatty Acids" -> "Fatty acids"
\item "Indoles Derivatives" -> "Indole derivatives"
\item "Nucleobases Related" -> "Nucleobase-related"
\item "Vitamins & Cofactors" -> "Vitamins and cofactors"
}
}
\examples{
metdat <- readBC(system.file("extdata", "DZHK_Serum_Feb2021_readBC.xlsx",
                             package="readBC"),
                 sheet = "Data Export")

}
